# Switch to NP from Jax to  improve attention manager performance