Current Jax implementation is slow. The operation array is tiny, leverage NP compute in CPU can boost the performance.    

The performance improved >30x. Below are two benchmark comparison before vs after this PR: 

**Before**
`
Total: 16 ms,  fill_new_page: 3ms, get_page_token_indices: 13ms
`

**After** 
`
Total: 0.5 ms, to_nparray: 0.1 ms, fill_new_page: 0.01ms,  get_page_token_indices: 0.05 ms to_jax_array:0.3 ms
`